refactored proc macro
