+++ sfsu-derive/src/hooks.rs
@@ -1,0 +1,54 @@
+use proc_macro2::{Ident, Span, TokenStream};
+use proc_macro_crate::{crate_name, FoundCrate};
+use proc_macro_error::abort_call_site;
+use quote::{quote, ToTokens};
+use syn::DeriveInput;
+
+pub fn hook_enum(input: DeriveInput) -> TokenStream {
+    let struct_name = {
+        let original_ident = &input.ident;
+        let og_ident_span = original_ident.span();
+        Ident::new(&format!("{}Raw", original_ident), og_ident_span)
+    };
+
+    let data = &input.data;
+
+    let mut variants = vec![];
+
+    match data {
+        syn::Data::Enum(ref e) => {
+            for v in &e.variants {
+                variants.push(v.ident.to_token_stream());
+            }
+        }
+        _ => abort_call_site!("Can only be derived for enums"),
+    };
+
+    let command_names = variants
+        .iter()
+        .map(|variant| heck::AsKebabCase(variant.to_string()).to_string())
+        .collect::<Vec<_>>();
+
+    let strum = match crate_name("strum").expect("strum is present in `Cargo.toml`") {
+        FoundCrate::Itself => Ident::new("strum", Span::call_site()),
+        FoundCrate::Name(name) => Ident::new(&name, Span::call_site()),
+    };
+
+    quote! {
+        // TODO: Better way of doing this? or add support for meta in proc macro
+        #[derive(Debug, Clone, #strum::Display, #strum::IntoStaticStr, #strum::EnumIter, PartialEq, Eq)]
+        #[strum(serialize_all = "kebab-case")]
+        pub enum #struct_name {
+            #(#variants),*
+        }
+
+        impl From<String> for #struct_name {
+            fn from(string: String) -> Self {
+                match string.as_str() {
+                    #(#command_names => #struct_name::#variants,)*
+                    _ => panic!("Invalid command name: {}", string),
+                }
+            }
+        }
+    }
+}
